stonks/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.8.1
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Flask==1.1.1
gast==0.3.3
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.18.1
pandas==1.0.0
postgres==3.0.0
protobuf==3.11.3
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
pyparsing==2.4.2
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.3
sqlparse==0.3.1
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
urllib3==1.25.3
virtualenv==16.7.0
Werkzeug==1.0.0
whitenoise==5.0.1
wrapt==1.12.1
